+++ src/configUI.ts
@@ -6,7 +6,7 @@ import type {
     ErmFirstCommaCommand,
     TConfigKey,
     TConfigs,
-    TDefProviderMethod,
+    TMethodMode,
 } from './configUI.data';
 import { EDiagMasterSwitch } from './configUI.data';
 import { statusBarItem } from './provider/vscWindows/statusBarItem';
@@ -38,7 +38,9 @@ function getConfig(Configs: vscode.WorkspaceConfiguration): TConfigs {
             showFileReport: getConfigs<boolean>(Configs, 'AhkNekoHelp.CodeLens.showFileReport'),
         },
         event: getConfigs<EFileRenameEvent>(Configs, 'AhkNekoHelp.event.FileRenameEvent'),
-        DefinitionProviderMethod: getConfigs<TDefProviderMethod>(Configs, 'AhkNekoHelp.DefinitionProvider.Method'),
+        method: {
+            gotoDef: getConfigs<TMethodMode>(Configs, 'AhkNekoHelp.method.gotoDef'),
+        },
         Diag: {
             AMasterSwitch: getConfigs<EDiagMasterSwitch>(Configs, 'AhkNekoHelp.Diag.AMasterSwitch'),
             code107: getConfigs<boolean>(Configs, 'AhkNekoHelp.Diag.code107LegacyAssignment'),
@@ -131,8 +133,8 @@ export function getCodeLenConfig(): TConfigs['CodeLens'] {
     return config.CodeLens;
 }
 
-export function getDefinitionProviderMethodConfig(): TConfigs['DefinitionProviderMethod'] {
-    return config.DefinitionProviderMethod;
+export function getMethodConfig(): TConfigs['method'] {
+    return config.method;
 }
 
 export function getFormatConfig(): TConfigs['format'] {