+++ src/provider/Def/DefProvider.ts
@@ -7,7 +7,6 @@ import { getDefSwitch } from './getDefSwitch';
 import { getDefWithLabel } from './getDefWithLabel';
 import { getFuncDef } from './getFuncDef';
 import { getMethodDef } from './getMethodDef';
-import { getThisMethod } from './getThisMethod';
 import { getValDefInFunc } from './getValDefInFunc';
 import { gotoIncludeDef } from './gotoIncludeDef';
 
@@ -21,11 +20,8 @@ function DefProviderCore(
     const IncludeFile: vscode.Location | null = gotoIncludeDef(AhkFileData, position);
     if (IncludeFile !== null) return [IncludeFile];
 
-    const methodDef: vscode.Location | null = getThisMethod(AhkFileData, position);
-    if (methodDef !== null) return [methodDef];
-
-    const methodDef2: vscode.Location[] | null = getMethodDef(document, position, AhkFileData);
-    if (methodDef2 !== null) return methodDef2;
+    const methodDef: vscode.Location[] | null = getMethodDef(document, position, AhkFileData);
+    if (methodDef !== null) return methodDef;
 
     const range: vscode.Range | undefined = document.getWordRangeAtPosition(
         position,